# specs/001-heliolus-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Heliolus Compliance Assessment Platform API
  description: API for compliance assessment, gap analysis, and vendor marketplace
  version: 1.0.0
  contact:
    name: Heliolus Support
    email: support@heliolus.com
servers:
  - url: https://api.heliolus.com/v1
    description: Production
  - url: https://staging-api.heliolus.com/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Development
security:
  - bearerAuth: []
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: List templates
      operationId: listTemplates
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/TemplateCategory'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
  /templates/{id}:
    get:
      tags: [Templates]
      summary: Get template
      operationId: getTemplate
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        404:
          $ref: '#/components/responses/NotFound'

  # Assessments
...
  /webhooks/stripe:
    post:
      tags: [Webhooks]
      summary: Stripe webhook
      operationId: stripeWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: Webhook processed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object
...
    Template:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/TemplateCategory'
        version:
          type: string
        isActive:
          type: boolean
        sections:
          type: array
          items:
            $ref: '#/components/schemas/TemplateSection'
    TemplateCategory:
      type: string
      enum: [FINANCIAL_CRIME, TRADE_COMPLIANCE, DATA_PRIVACY, CYBERSECURITY, ESG]
    TemplateSection:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/TemplateQuestion'
        weight:
          type: number
        order:
          type: integer
    TemplateQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        type:
          $ref: '#/components/schemas/QuestionType'
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
        helpText:
          type: string
        order:
          type: integer
    QuestionType:
      type: string
      enum: [TEXT, NUMBER, SELECT, MULTISELECT, BOOLEAN, FILE, DATE]
...